Flexbox to the rescue!
Switched so that when in portrait orientation it defaults to 90vw,
which means it’ll always be in screen. Flexbox also perfectly centers
it on all screens vertically (aka CSS magic).

I also changed the zoom on low-res screens (for iPhone resolution
spoofing) to 0.5, so it’ll fit in screen when horizontal as well.